fix gameserver crash when increasing max_players. Various other fixes
gameserver:
don't use max_players to iterate player list as it may change. Use
static size array instead (same as lobby).
fixes crash due to uninitialized data after increasing max_players.

waitmodule:
trim white spaces around username.

lobby:
only SDO needs a pipe to the lobby.
simplify ping.
ignore bogus LEAVESESSION messages.

fix warnings

+++ gs_lobby/discord.c
@@ -50,7 +50,7 @@ static int writeJsonString(char *json, const char *s)
   *j++ = '"';
   *j++ = '\0';
 
-  return j - json - 1;
+  return (int)(j - json - 1);
 }
 
 static void freeNotif(Notif *notif)
